+++ SharpSource/SharpSource.Test/TestMethodWithoutTestAttributeTests.cs
@@ -205,5 +205,20 @@ namespace ConsoleApplication1
 
             VerifyDiagnostic(original);
         }
+
+        [BugVerificationTest(IssueUrl = "https://github.com/Vannevelj/SharpSource/issues/91")]
+        [DataRow("record")]
+        [DataRow("record class")]
+        [DataRow("record struct")]
+        public void TestMethodWithoutTestAttribute_Record(string record)
+        {
+            var original = $@"
+{record} Test
+{{
+    public void MyMethod() {{ }}
+}}";
+
+            VerifyDiagnostic(original);
+        }
     }
 }